+++ system/src/Grav/Common/Assets/JsModule.php
@@ -12,21 +12,21 @@ namespace Grav\Common\Assets;
 use Grav\Common\Utils;
 
 /**
- * Class InlineCss
+ * Class Js
  * @package Grav\Common\Assets
  */
-class InlineCss extends BaseAsset
+class JsModule extends BaseAsset
 {
     /**
-     * InlineCss constructor.
+     * Js constructor.
      * @param array $elements
      * @param string|null $key
      */
     public function __construct(array $elements = [], ?string $key = null)
     {
         $base_options = [
-            'asset_type' => 'css',
-            'position' => 'after'
+            'asset_type' => 'js_module',
+            'attributes' => ['type' => 'module']
         ];
 
         $merged_attributes = Utils::arrayMergeRecursiveUnique($base_options, $elements);
@@ -34,11 +34,16 @@ class InlineCss extends BaseAsset
         parent::__construct($merged_attributes, $key);
     }
 
-    /**
+        /**
      * @return string
      */
     public function render()
     {
-        return '<style' . $this->renderAttributes(). ">\n" . trim($this->asset) . "\n</style>\n";
+        if (isset($this->attributes['loading']) && $this->attributes['loading'] === 'inline') {
+            $buffer = $this->gatherLinks([$this], self::JS_MODULE_ASSET);
+            return '<script' . $this->renderAttributes() . ">\n" . trim($buffer) . "\n</script>\n";
+        }
+
+        return '<script src="' . trim($this->asset) . $this->renderQueryString() . '"' . $this->renderAttributes() . $this->integrityHash($this->asset) . "></script>\n";
     }
 }