+++ system/src/Grav/Common/Media/Interfaces/MediaObjectInterface.php
@@ -64,6 +64,14 @@ interface MediaObjectInterface extends \Grav\Framework\Media\Interfaces\MediaObj
     public function addAlternative($ratio, MediaObjectInterface $alternative);
 
     /**
+     * Get list of image alternatives. Includes the current media image as well.
+     *
+     * @param bool $withDerived If true, include generated images as well. If false, only return existing files.
+     * @return array
+     */
+    public function getAlternatives(bool $withDerived = true): array;
+
+    /**
      * Return string representation of the object (html).
      *
      * @return string