+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1. [](#improved)
     * Improved `Utils::download()` method to allow overrides on download name, mime and expires header
+    * Improved `onPageFallBackUrl` event
 3. [](#bugfix)
     * Fixed CLI `--env` and `--lang` options having no effect if they aren't added before all the other options