+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.7.27
-## 01/04/2022
+## mm/dd/2022
 
 1. [](#improved)
     * Improved `Utils::download()` method to allow overrides on download name, mime and expires header