+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 1. [](#new)
    * Added support for generic `assets.link()` for external references. No pipeline support.
-   * Added support for `assets.addJSmodule()` with full pipeline support.
+   * Added support for `assets.addJsModule()` with full pipeline support.
 2. [](#improved)
     * Improved `Utils::download()` method to allow overrides on download name, mime and expires header
     * Improved `onPageFallBackUrl` event